add go-to-match (pretty cool)
Just to note, this change adds actually the first svn invocation to the
script - svn list (they have been in the configuration only). So it
turns out that the script could be used to browse whatever; but it is
intended to be used with Subversion (making the script actually useful),
and the change allows to add a helpful feature in a reasonable way.